Add fragment_login landscape. Update wiki
