+++ app/SwimChrono/app/src/main/res/layout-land/fragment_login.xml
@@ -7,7 +7,7 @@
 
     <EditText
         android:id="@+id/email_edit_text"
-        android:layout_width="match_parent"
+        android:layout_width="400dp"
         android:layout_height="wrap_content"
         android:autofillHints="email"
         android:hint="@string/example_email"
@@ -19,7 +19,7 @@
 
     <EditText
         android:id="@+id/password_edit_text"
-        android:layout_width="match_parent"
+        android:layout_width="400dp"
         android:layout_height="wrap_content"
         android:autofillHints=""
         android:hint="@string/password"
@@ -32,8 +32,8 @@
         android:id="@+id/login_button"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="104dp"
         android:text="@string/login"
+        app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.498"
         app:layout_constraintStart_toStartOf="parent"
