+++ docs/user-guide/tools.md
@@ -4,6 +4,3 @@
 
 - [**Mkdocs**](https://www.mkdocs.org/): Herramienta para la creación de sitios web de documentación.
 
-- [**FastApi**](https://fastapi.tiangolo.com/): Framework que nos permite crear un backend donde tener todos los datos de los nadadores.
-
-- [**Render.com**](https://render.com/): Plataforma en la nube que ofrece servicios de alojamiento y ofrece despliegues de aplicaciones que podrán ser consumidas por la aplicación
